refactor: extract utility functions for Highcharts configuration into separate file and simplify chart options handling
